Refactor precision in analytics, performance, manufacturing, and SCM modules

+++ src/analytics.rs
@@ -6,10 +6,10 @@ use serde::{Deserialize, Serialize};
 pub struct AnalyticsModel {
     pub model_id: String,
     pub model_type: String,
-    pub accuracy: f64,
-    pub precision: f64,
-    pub recall: f64,
-    pub f1_score: f64,
+    pub accuracy: f32,
+    pub precision: f32,
+    pub recall: f32,
+    pub f1_score: f32,
 }
 
 #[napi]
@@ -18,10 +18,10 @@ pub fn calculate_model_accuracy(
     true_negatives: i32,
     false_positives: i32,
     false_negatives: i32,
-) -> f64 {
+) -> f32 {
     let total = true_positives + true_negatives + false_positives + false_negatives;
     if total > 0 {
-        ((true_positives + true_negatives) as f64 / total as f64) * 100.0
+        ((true_positives + true_negatives) as f32 / total as f32) * 100.0
     } else {
         0.0
     }
@@ -29,21 +29,21 @@ pub fn calculate_model_accuracy(
 
 #[napi]
 pub fn calculate_regression_metrics(
-    actual_values: Vec<f64>,
-    predicted_values: Vec<f64>,
-) -> f64 {
+    actual_values: Vec<f32>,
+    predicted_values: Vec<f32>,
+) -> f32 {
     if actual_values.len() != predicted_values.len() || actual_values.is_empty() {
         return 0.0;
     }
 
-    let mean_actual = actual_values.iter().sum::<f64>() / actual_values.len() as f64;
+    let mean_actual = actual_values.iter().sum::<f32>() / actual_values.len() as f32;
     
-    let ss_res: f64 = actual_values.iter()
+    let ss_res: f32 = actual_values.iter()
         .zip(predicted_values.iter())
         .map(|(actual, predicted)| (actual - predicted).powi(2))
         .sum();
     
-    let ss_tot: f64 = actual_values.iter()
+    let ss_tot: f32 = actual_values.iter()
         .map(|actual| (actual - mean_actual).powi(2))
         .sum();
     
@@ -56,14 +56,14 @@ pub fn calculate_regression_metrics(
 
 #[napi]
 pub fn calculate_predictive_accuracy(
-    predictions: Vec<f64>,
-    actuals: Vec<f64>,
-) -> f64 {
+    predictions: Vec<f32>,
+    actuals: Vec<f32>,
+) -> f32 {
     if predictions.len() != actuals.len() || predictions.is_empty() {
         return 0.0;
     }
 
-    let mape: f64 = predictions.iter()
+    let mape: f32 = predictions.iter()
         .zip(actuals.iter())
         .map(|(pred, actual)| {
             if *actual != 0.0 {
@@ -72,7 +72,7 @@ pub fn calculate_predictive_accuracy(
                 0.0
             }
         })
-        .sum::<f64>() / predictions.len() as f64;
+        .sum::<f32>() / predictions.len() as f32;
 
     (1.0 - mape) * 100.0
 }
